move: field 폴더 컴포넌트 폴더로 이동 및 이름 변경
